Refactor the way functions are exported.
